add asp.net core miniaml web api with customer endpoint to northwind db

+++ textbooks/cs14net10/ch12/260326/TPModernWeb/TP.EntityModels.Sqlite/Order.cs
@@ -38,7 +38,8 @@ public partial class Order
     public int? ShipVia { get; set; }
 
     [Column(TypeName = "money")]
-    public double? Freight { get; set; }
+    //public double? Freight { get; set; }
+    public decimal? Freight { get; set; }
 
     [Column(TypeName = "nvarchar (40)")]
     public string? ShipName { get; set; }
